zip/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Course-manager"
version = "1.0.0"
description = "Course Management System with Learning Outcome Tracking"
readme = "README.md"
authors = [
    {name = "Dr. Kerem ERCOŞKUN", email = "ercoskunkerem@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Flask",
    "Topic :: Education",
]
requires-python = ">=3.8"
dependencies = [
    "Flask>=2.3.0",
    "Flask-WTF>=1.1.0",
    "SQLAlchemy>=2.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "Jinja2>=3.1.2",
    "Werkzeug>=2.3.0",
    "markdown>=3.4.4",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
    "typing-extensions>=4.5.0",
    "xhtml2pdf>=0.2.13",
    "reportlab>=4.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "coverage>=7.0.0",
    "mypy>=1.5.0",
    "pylint>=2.17.0",
]
api = [
    "Flask-RESTful>=0.3.10",
    "Flask-CORS>=4.0.0",
]
reports = [
    "weasyprint>=58.0",
    "pdfkit>=1.0.0",
    "markdown2>=2.4.0",
]
interop = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/keremer/pearson"
Documentation = "https://github.com/keremer/pearson/wiki"

[project.scripts]
# Updated to point to the main module within the pearson package
crminaec = "crminaec.main:cli"

[tool.setuptools.packages.find]
where = ["."]
# Focus on the 'crminaec' root package; setuptools will find the 
# sub-packages (web, api, cli, interop) automatically if they contain __init__.py
include = ["crminaec*"]
exclude = ["tests*", "*.tests", "*.tests.*", "tests.*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true